+++ src/resources/views/products/create-variant.blade.php
@@ -1,0 +1,124 @@
+@extends('redminportal::layouts.bare')
+
+@section('content')
+    
+    @include('redminportal::partials.errors')
+
+    {!! Form::open(array('files' => TRUE, 'action' => '\Redooor\Redminportal\App\Http\Controllers\ProductController@postStore', 'role' => 'form')) !!}
+
+    {!! Form::hidden('product_id', $product_id, array('id' => 'product_id')) !!}
+
+    	<div class='row'>
+            <div class="col-sm-4 col-sm-push-8">
+                <div class="well">
+                    <div class='form-actions text-right'>
+                        {!! Form::submit(Lang::get('redminportal::buttons.create'), array('class' => 'btn btn-primary btn-sm')) !!}
+                    </div>
+                </div>
+                <div class='well well-small'>
+                    <div class="form-group">
+                        <div class="checkbox">
+                            <label for="featured-checker">
+                                {!! Form::checkbox('featured', true, true, array('id' => 'featured-checker')) !!} {{ Lang::get('redminportal::forms.featured') }}
+                            </label>
+                        </div>
+                    </div>
+
+                    <div class="form-group">
+                        <div class="checkbox">
+                            <label for="active-checker">
+                                {!! Form::checkbox('active', true, true, array('id' => 'active-checker')) !!} {{ Lang::get('redminportal::forms.active') }}
+                            </label>
+                        </div>
+                    </div>
+                </div>
+                <div class="panel panel-default">
+                    <div class="panel-heading">
+                        <div class="panel-title">{{ Lang::get('redminportal::forms.category') }}</div>
+                    </div>
+                    <div class="panel-body">
+                        {!! Form::hidden('category_id', null, array('id' => 'category_id')) !!}
+                        <ul class="redooor-hierarchy">
+                        @foreach ($categories as $item)
+                            <li>{!! $item->printCategory() !!}</li>
+                        @endforeach
+                        </ul>
+                    </div>
+                </div>
+		        <div>
+                    <div class="fileupload fileupload-new" data-provides="fileupload">
+                      <div class="fileupload-preview thumbnail" style="width: 200px; height: 150px;"></div>
+                      <div>
+                        <span class="btn btn-default btn-file"><span class="fileupload-new">{{ Lang::get('redminportal::forms.select_image') }}</span><span class="fileupload-exists">{{ Lang::get('redminportal::forms.change_image') }}</span>{!! Form::file('image') !!}</span>
+                        <a href="#" class="btn btn-danger fileupload-exists" data-dismiss="fileupload">{{ Lang::get('redminportal::forms.remove_image') }}</a>
+                      </div>
+                    </div>
+                </div>
+	        </div>
+
+	        <div class="col-sm-8 col-sm-pull-4">
+                <div class="panel panel-default">
+                    <div class="panel-heading">
+                        <h4 class="panel-title">{{ Lang::get('redminportal::forms.create_product') }}</h4>
+                    </div>
+                    <div class="panel-body">
+                        @include('redminportal::partials.lang-selector', ['selector_name' => '-variant'])
+                    </div>
+                </div>
+                <div class="panel panel-default">
+                    <div class="panel-heading">
+                        <h4 class="panel-title">{{ Lang::get('redminportal::forms.product_properties') }}</h4>
+                    </div>
+                    <div class="panel-body">
+                        <div class="form-group">
+                            {!! Form::label('sku', Lang::get('redminportal::forms.sku')) !!}
+                            {!! Form::text('sku', null, array('class' => 'form-control')) !!}
+                        </div>
+                        <div class="form-group">
+                            {!! Form::label('price', Lang::get('redminportal::forms.price')) !!}
+                            <div class="input-group">
+                                <span class="input-group-addon">$</span>
+                                {!! Form::text('price', null, array('class' => 'form-control')) !!}
+                            </div>
+                        </div>
+                        <div class="form-group">
+                            {!! Form::label('tags', Lang::get('redminportal::forms.tags_separated_by_comma')) !!}
+                            {!! Form::text('tags', null, array('class' => 'form-control')) !!}
+                        </div>
+                    </div>
+                </div>
+	        </div>
+        </div>
+    {!! Form::close() !!}
+@stop
+
+@section('footer')
+    <script src="{{ URL::to('vendor/redooor/redminportal/js/bootstrap-fileupload.js') }}"></script>
+    <script>
+        !function ($) {
+            $(function(){
+                // On load, check if previous category exists for error message
+                function checkCategory() {
+                    $selected_val = $('#category_id').val();
+                    if ($selected_val != '') {
+                        $('.redooor-hierarchy a').each(function() {
+                            if ($(this).attr('href') == $selected_val) {
+                                $(this).addClass('active');
+                            }
+                        });
+                    }
+                }
+                checkCategory();
+                // Change selected category
+                $(document).on('click', '.redooor-hierarchy a', function(e) {
+                    e.preventDefault();
+                    $selected = $(this).attr('href');
+                    $('#category_id').val($selected);
+                    $('.redooor-hierarchy a.active').removeClass('active');
+                    $(this).addClass('active');
+                });
+            })
+        }(window.jQuery);
+    </script>
+    @include('redminportal::plugins/tinymce')
+@stop